factor out FixStdioStreams
