+++ shell/app/uv_stdio_fix.h
@@ -1,0 +1,10 @@
+// Copyright (c) 2022 Slack Technologies, Inc.
+// Use of this source code is governed by the MIT license that can be
+// found in the LICENSE file.
+
+#ifndef ELECTRON_SHELL_APP_UV_STDIO_FIX_H_
+#define ELECTRON_SHELL_APP_UV_STDIO_FIX_H_
+
+void FixStdioStreams();
+
+#endif  // ELECTRON_SHELL_APP_UV_STDIO_FIX_H_
